# backend/private/openapi.yaml
openapi: 3.0.0
info:
  title: MOCK Web Server
  version: 2.0.0
  description: Versión NodeJS del **servidor de pruebas** para cursos de FrontEnd.
  contact:
    name: Javier Martín
    url: https://github.com/jmagit
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: "{protocol}://localhost:{port}/"
    description: Servidor local para las pruebas
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      port:
        enum:
          - "4321"
          - "8181"
        default: "4321"
...
paths:
  /api/contactos:
    get:
      tags:
        - contactos
      summary: Listar contactos
      parameters:
        - $ref: "#/components/parameters/pagina"
        - $ref: "#/components/parameters/filas"
        - $ref: "#/components/parameters/ordenar"
        - $ref: "#/components/parameters/buscar"
        - $ref: "#/components/parameters/proyeccion"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Contactos"
                  - $ref: "#/components/schemas/ContactosPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags:
        - contactos
      summary: Crear contacto
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactoModel"
        required: true
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    options:
      tags:
        - contactos
      summary: Sondeo CORS
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  "/api/contactos/{id}":
    parameters:
      - name: id
        in: path
        required: true
        description: Identificador del recurso
        schema:
          type: integer
    get:
      tags:
        - contactos
      summary: Recuperar contacto
      parameters:
        - in: query
          name: _projection
          description: Devuelve solo aquellas propiedades de la lista suministrada, los
            nombres de las propiedades deben ir separadas por comas.
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactoProjection"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - contactos
      summary: Reemplazar contacto
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactoModel"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactoProjection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    options:
      tags:
        - contactos
      summary: Sondeo CORS
      responses:
        "200":
          description: OK
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      tags:
        - contactos
      summary: Actualizar parcialmente contacto
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactoProjection"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactoProjection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - contactos
      summary: Borrar contacto
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  schemas:
    ErrorMessage:
      type: object
      title: Error message
      required:
        - type
        - status
        - title
      properties:
        type:
          type: string
        status:
          type: integer
        title:
          type: string
        detail:
          type: string
        source:
          type: string
    ContactoModel:
      title: Contacto (Modelo)
      allOf:
        - $ref: "#/components/schemas/ContactoProjection"
      required:
        - id
        - nombre
    ContactoProjection:
      title: Contacto (Proyección)
      type: object
      properties:
        id:
          title: Id
          type: integer
          example: 1
          description: El 0 actúa como autonumérico en la creación
        tratamiento:
          title: Tratamiento
          type: string
          enum:
            - Sr.
            - Sra.
            - Srta.
            - Dr.
            - Dra.
            - Ilmo.
            - Ilma.
            - Excmo.
            - Excma.
          example: Excmo.
        nombre:
          title: Nombre
          type: string
          maxLength: 100
          example: Pietro
        apellidos:
          title: Apellidos
          type: string
          maxLength: 100
          example: Woodall
        telefono:
          title: Telefono
          type: string
          example: 642 051 348
          pattern: ^(\d{3}\s){2}\d{3}$
        email:
          title: Email
          type: string
          format: email
          maxLength: 100
          example: pwoodall0@army.mil
        nif:
          title: N.I.F.
          type: string
          format: nif
          example: 12345678Z
        sexo:
          title: Sexo
          type: string
          enum:
            - H
            - M
          example: H
        nacimiento:
          title: Nacimiento
          type: string
          format: date
          example: 1963-10-11
        avatar:
          title: Avatar
          type: string
          format: uri
          maxLength: 200
          example: https://randomuser.me/api/portraits/men/0.jpg
        conflictivo:
          title: Conflictivo
          type: boolean
          example: true
        icono:
          title: Icono
          type: string
          maxLength: 65000
          example: data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABGdBTUEAAK/INwWK6QAAABl0RVh0U29mdHdhcmUAQWRvYmUgSW1hZ2VSZWFkeXHJZTwAAALJSURBVDjLpVNbS1RRFP72nHGaUedmpamj5pgKSUli0eXJfOipoELs+hA+FFT44oMSRBkZ1GNC/oAoCJ+C8QaSZhdCTEU0rbyP05jTzOTczv20zkz6IBFBBxZrs9f+vv3t76zFNE3D/3zGjYXa01OuKUqtJstZqizPUnSb6+vDem29o8OpiuIFimZFEOZUQXiY09rq0WtMV6B2d98k8F2YTU4YOKjRCKTw+qQqSZcJFKPoTMty7Nu2pwRKJILY2DjiPv9TIrvGlK6uk5osvjSUl4ERGJIEmEyQlpeRmJpeUkQxaikp3mt2u4FwEPoFcDgR6u1FaHzivpGk3uDy88DiccDnBxQVkGWkFReAO3SgUONFcGYLMDxC71RTF1jTYTt2FIEPw1eMJLOCs9mAuXkqyoBGhxQF+PQFBjvtcwyY+aq7BIhyKgeCYCUqFJ636QSmJIAXAEFIEYhSimTFl1STvFkPjfYEqqmUJRFkpmokg9bJ+Z1MpgKfIICSAutAZcs6mSmMBvLCADLRYCCWfnlmBnDlAbEokCASnqcbhFQW+VRO7lPQX0CpG4GB17qCZyzc3u4mFf2WItdu864cYHIKWAv+bg5lU4V3bYxiFLHQd3CZNqT9zIynx7PLuAceTyg2OPiK9/kPqqEfeWmV+8HSqb8WvSkwue5dHUXIHkbFmQaUXmxETrYVIX6FC0YmFtlGK39rabHTm5pIVqOzqtJqzc8F3r2nNoxiaLkT1Q1NsMwOkLFvALsDAWMRPvYNzbOtszDf0FBNRI8dxYWHd9TWJE3rbzuF449egFWd3jwXvpOLkaGAxv40TJ/r6hxEco+8ue4syGVTC88TR85fsmQseSAk/KCWQ2Sdw/QsVtjfpnGipuYs9Ykr4pp3mqwZtwq2y0ajwYvImoyFVU6ReO02+9dxfnuusDkW9F3lFFakcNoKoZ6c6JPbfgHHGqU/+iLy1wAAAABJRU5ErkJggg==
        direcciones:
          title: Direcciones
          type: array
          uniqueItems: true
          items:
            title: Dirección
            type: object
            required:
              - calle
              - codigo
              - provincia
              - pais
            properties:
              calle:
                title: Calle
                type: string
                maxLength: 500
                example: 488 Pankratz Junction
              codigo:
                title: Código Postal
                type: string
                example: "29071"
                pattern: ^\d{4,5}$
              provincia:
                title: Provincia
                type: string
                maxLength: 100
                example: Malaga
              pais:
                title: Pais
                type: string
                maxLength: 100
                example: España
      additionalProperties: false
      minProperties: 1
    Contactos:
      type: array
      title: Contactos (Listado)
      items:
        $ref: "#/components/schemas/ContactoProjection"
    ContactosPage:
      type: object
      title: Contactos (Página)
      properties:
        content:
          $ref: "#/components/schemas/Contactos"
        totalElements:
          type: integer
          description: Número total de elementos
        totalPages:
          type: integer
          description: Número total de páginas
        size:
          type: integer
          description: Tamaño de página en elementos
        number:
          type: integer
          description: Número de página actual
        numberOfElements:
          type: integer
          description: Número total de elementos en la página
        empty:
          type: boolean
          description: Si la página está vacía
        first:
          type: boolean
          description: Si la página es la primera
        last:
          type: boolean
          description: Si la página es la última
...
  responses:
    Created:
      description: Created
      headers:
        location:
          description: URL al elemento recién creado
          schema:
            type: string
            format: uri
    NoContent:
      description: No content
    BadRequest:
      description: Invalid data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
  parameters:
    pagina:
      in: query
      name: _page
      description: Número de página empezando en 0 (primera página).
      required: false
      schema:
        oneOf:
          - type: integer
            minimum: 0
          - type: string
            enum:
              - count
              - COUNT
    filas:
      in: query
      name: _rows
      description: Número de filas por página, por defecto 20 si se omite pero aparece
        el parámetro *_page*.
      required: false
      schema:
        type: integer
        minimum: 0
    ordenar:
      in: query
      name: _sort
      description: Indica la lista de propiedades (separadas por comas) por la que se
        ordenaran los resultados, en caso de omitirse se utilizará la propiedad
        que actúa como primary key. Si el nombre de la propiedad está precedido
        por un guion (signo negativo) la ordenación será descendente.
      required: false
      allowReserved: true
      schema:
        type: string
    buscar:
      in: query
      name: _search
      description: Selecciona todos aquellos que en alguna de sus propiedades contenga
        el valor proporcionado.
      required: false
      allowReserved: true
      schema:
        type: string
    proyeccion:
      in: query
      name: _projection
      description: Devuelve solo aquellas propiedades de la lista suministrada, los
        nombres de las propiedades deben ir separadas por comas.
      required: false
      allowReserved: true
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: Authorization